Move get_linker to linker.rs
